rolled up ThemeOption into Theme.cs
